requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.7.0
awscli==1.22.66
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
boto3==1.21.11
botocore==1.24.11
brotlipy==0.7.0
cachetools==5.2.0
calmsize==0.1.3
ccimport==0.3.7
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
constantly==15.1.0
cryptography==3.4.7
cumm-cu111==0.2.8
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
docutils==0.15.2
easydict==1.9
executing==0.8.3
fairscale==0.4.12
fire==0.4.0
flatbuffers==2.0
frozenlist==1.3.1
fsspec==2022.7.1
funcsigs==1.0.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
hkdf==0.0.3
humanize==3.6.0
hyperlink==21.0.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.21.1
importlib-metadata==4.12.0
imutils==0.5.4
incremental==21.3.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.11.1
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
Keras-Preprocessing==1.1.2
lark==1.1.2
laspy==2.1.2
libclang==13.0.0
llvmlite==0.38.1
magic-wormhole==0.12.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.2
nbclassic==0.3.6
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8.5
ninja==1.10.2.3
notebook==6.4.8
notebook-shim==0.1.0
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
open3d==0.13.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pccm==0.3.4
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.5.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-goicp==0.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycodestyle==2.9.1
pycparser==2.21
pyDeprecate==0.3.2
pymongo==4.0.1
PyNaCl==1.4.0
pyntcloud==0.1.6
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.6.5
pytorch-memlab==0.2.4
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.2
scale-lidar-io==0.1.163
scale-lidar-io-debug==0.1.3
scaleapi==2.6.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.9.0
Send2Trash==1.8.0
sentry-sdk==1.9.3
service-identity==21.1.0
setproctitle==1.3.1
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
spake2==0.8
-e git+https://github.com/facebookresearch/SparseConvNet.git@dcf6a7ff540e1825ffe48ba6b2c1493ba18788b2#egg=sparseconvnet
spconv-cu111==2.1.21
stack-data==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.2
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.8.8
tldr==2.0.0
toml==0.10.2
tomlkit==0.10.0
torch==1.10.1+cu111
torch-scatter @ file:///home2/mznv82/cvav/torch_scatter-2.0.9-cp38-cp38-linux_x86_64.whl
torchaudio==0.10.1+rocm4.1
torchfile==0.1.0
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.11.2+cu111
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
Twisted==21.2.0
txaio==21.2.1
txtorcon==20.0.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
ujson==5.1.0
urllib3==1.26.11
visdom==0.1.8.9
wandb==0.13.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.8.1
zipp==3.8.1
zope.interface==5.4.0
